pyproject.toml
[tool.poetry]
name = "gsay"
version = "0.1.3"
description = "Nightcored Google TTS"
authors = ["Johannes C. Mayer <j.c.mayer240@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.scripts]
gsay = "gsay:main"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0.1"
google-cloud-texttospeech = "^2.15.0"
xdg-base-dirs = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
